+++ test/support/test_versioned_workflows.ex
@@ -1,5 +1,5 @@
 defmodule Hephaestus.Test.Versioned.V1 do
-  use Hephaestus.Workflow, version: 1
+  use Hephaestus.Workflow, version: 1, unique: [key: "testversioned"]
 
   @impl true
   def start, do: Hephaestus.Test.V2.StepA
@@ -9,7 +9,7 @@ defmodule Hephaestus.Test.Versioned.V1 do
 end
 
 defmodule Hephaestus.Test.Versioned.V2 do
-  use Hephaestus.Workflow, version: 2
+  use Hephaestus.Workflow, version: 2, unique: [key: "testversioned"]
 
   @impl true
   def start, do: Hephaestus.Test.V2.StepA
@@ -21,11 +21,12 @@ end
 defmodule Hephaestus.Test.Versioned do
   use Hephaestus.Workflow,
     versions: %{1 => __MODULE__.V1, 2 => __MODULE__.V2},
-    current: 2
+    current: 2,
+    unique: [key: "testversioned"]
 end
 
 defmodule Hephaestus.Test.VersionedWithCallback.V1 do
-  use Hephaestus.Workflow, version: 1
+  use Hephaestus.Workflow, version: 1, unique: [key: "testversioned"]
 
   @impl true
   def start, do: Hephaestus.Test.V2.StepA
@@ -35,7 +36,7 @@ defmodule Hephaestus.Test.VersionedWithCallback.V1 do
 end
 
 defmodule Hephaestus.Test.VersionedWithCallback.V2 do
-  use Hephaestus.Workflow, version: 2
+  use Hephaestus.Workflow, version: 2, unique: [key: "testversioned"]
 
   @impl true
   def start, do: Hephaestus.Test.V2.StepA
@@ -50,7 +51,10 @@ defmodule Hephaestus.Test.VersionedWithCallback do
       1 => __MODULE__.V1,
       2 => __MODULE__.V2
     },
-    current: 2
+    current: 2,
+    unique: [key: "testversionedcb"]
 
-  def version_for(%{1 => _}, force_v1: true), do: 1
+  def version_for(%{1 => _}, opts) do
+    if opts[:force_v1], do: 1
+  end
 end