+++ test/support/test_workflows.ex
@@ -1,5 +1,5 @@
 defmodule Hephaestus.Test.LinearWorkflow do
-  use Hephaestus.Workflow
+  use Hephaestus.Workflow, unique: [key: "test"]
 
   @impl true
   def start, do: Hephaestus.Test.Linear.StepA
@@ -10,7 +10,7 @@ defmodule Hephaestus.Test.LinearWorkflow do
 end
 
 defmodule Hephaestus.Test.BranchWorkflow do
-  use Hephaestus.Workflow
+  use Hephaestus.Workflow, unique: [key: "test"]
 
   @impl true
   def start, do: Hephaestus.Test.BranchStep
@@ -23,14 +23,18 @@ defmodule Hephaestus.Test.BranchWorkflow do
 end
 
 defmodule Hephaestus.Test.ParallelWorkflow do
-  use Hephaestus.Workflow
+  use Hephaestus.Workflow, unique: [key: "test"]
 
   @impl true
   def start, do: Hephaestus.Test.Parallel.Start
 
   @impl true
   def transit(Hephaestus.Test.Parallel.Start, :done, _ctx),
-    do: [Hephaestus.Test.Parallel.BranchA, Hephaestus.Test.Parallel.BranchB, Hephaestus.Test.Parallel.BranchC]
+    do: [
+      Hephaestus.Test.Parallel.BranchA,
+      Hephaestus.Test.Parallel.BranchB,
+      Hephaestus.Test.Parallel.BranchC
+    ]
 
   def transit(Hephaestus.Test.Parallel.BranchA, :done, _ctx), do: Hephaestus.Test.Parallel.Join
   def transit(Hephaestus.Test.Parallel.BranchB, :done, _ctx), do: Hephaestus.Test.Parallel.Join
@@ -39,7 +43,7 @@ defmodule Hephaestus.Test.ParallelWorkflow do
 end
 
 defmodule Hephaestus.Test.MixedParallelWorkflow do
-  use Hephaestus.Workflow
+  use Hephaestus.Workflow, unique: [key: "test"]
 
   @impl true
   def start, do: Hephaestus.Test.MixedParallel.Start
@@ -48,13 +52,17 @@ defmodule Hephaestus.Test.MixedParallelWorkflow do
   def transit(Hephaestus.Test.MixedParallel.Start, :done, _ctx),
     do: [Hephaestus.Test.MixedParallel.Sync, Hephaestus.Test.MixedParallel.Async]
 
-  def transit(Hephaestus.Test.MixedParallel.Sync, :done, _ctx), do: Hephaestus.Test.MixedParallel.Join
-  def transit(Hephaestus.Test.MixedParallel.Async, :timeout, _ctx), do: Hephaestus.Test.MixedParallel.Join
+  def transit(Hephaestus.Test.MixedParallel.Sync, :done, _ctx),
+    do: Hephaestus.Test.MixedParallel.Join
+
+  def transit(Hephaestus.Test.MixedParallel.Async, :timeout, _ctx),
+    do: Hephaestus.Test.MixedParallel.Join
+
   def transit(Hephaestus.Test.MixedParallel.Join, :done, _ctx), do: Hephaestus.Steps.Done
 end
 
 defmodule Hephaestus.Test.AsyncWorkflow do
-  use Hephaestus.Workflow
+  use Hephaestus.Workflow, unique: [key: "test"]
 
   @impl true
   def start, do: Hephaestus.Test.Async.StepA
@@ -66,13 +74,16 @@ defmodule Hephaestus.Test.AsyncWorkflow do
 end
 
 defmodule Hephaestus.Test.EventWorkflow do
-  use Hephaestus.Workflow
+  use Hephaestus.Workflow, unique: [key: "test"]
 
   @impl true
   def start, do: Hephaestus.Test.Event.StepA
 
   @impl true
   def transit(Hephaestus.Test.Event.StepA, :done, _ctx), do: Hephaestus.Test.Event.WaitForEvent
-  def transit(Hephaestus.Test.Event.WaitForEvent, :payment_confirmed, _ctx), do: Hephaestus.Test.Event.StepB
+
+  def transit(Hephaestus.Test.Event.WaitForEvent, :payment_confirmed, _ctx),
+    do: Hephaestus.Test.Event.StepB
+
   def transit(Hephaestus.Test.Event.StepB, :done, _ctx), do: Hephaestus.Steps.Done
 end